requirements.txt
# Farm Game Bot - Requirements

# Telegram Bot API
pyTelegramBotAPI==4.14.0

# Environment variables
python-dotenv==1.0.0
